Trac 2910 Reverting the "DOUBLE" problem: NetCDF v3 doesn't support LONGs.

+++ atlas-data-storage/src/main/java/uk/ac/ebi/gxa/data/NetCDFDataCreator.java
@@ -242,7 +242,7 @@ public class NetCDFDataCreator {
                 new Dimension[]{designElementDimension, designElementLenDimension}
         );
         netCdf.addVariable(
-                "GN", DataType.LONG,
+                "GN", DataType.DOUBLE,
                 new Dimension[]{designElementDimension}
         );
